fix for import able on eHelper

+++ src/main/java/io/owslab/mailreceiver/service/expansion/BusinessPartnerService.java
@@ -1,5 +1,6 @@
 package io.owslab.mailreceiver.service.expansion;
 
+import io.owslab.mailreceiver.controller.PartnerImportExportController;
 import io.owslab.mailreceiver.dao.BusinessPartnerDAO;
 import io.owslab.mailreceiver.dao.BusinessPartnerGroupDAO;
 import io.owslab.mailreceiver.dto.CSVPartnerDTO;
@@ -9,9 +10,13 @@ import io.owslab.mailreceiver.exception.PartnerCodeException;
 import io.owslab.mailreceiver.form.PartnerForm;
 import io.owslab.mailreceiver.model.BusinessPartner;
 import io.owslab.mailreceiver.model.BusinessPartnerGroup;
+import io.owslab.mailreceiver.service.file.UploadFileService;
 import io.owslab.mailreceiver.utils.CSVBundle;
 import io.owslab.mailreceiver.utils.Utils;
+import org.apache.commons.lang.exception.ExceptionUtils;
 import org.mozilla.universalchardet.UniversalDetector;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
@@ -31,11 +36,16 @@ import java.util.*;
  */
 @Service
 public class BusinessPartnerService {
+    private static final Logger logger = LoggerFactory.getLogger(BusinessPartnerService.class);
+
     @Autowired
     private BusinessPartnerDAO partnerDAO;
 
     @Autowired
     private BusinessPartnerGroupDAO partnerGroupDAO;
+
+    @Autowired
+    private UploadFileService uploadFileService;
 
     public List<BusinessPartner> getAll() {
         return (List<BusinessPartner>) partnerDAO.findAll();
@@ -178,12 +188,11 @@ public class BusinessPartnerService {
 
     public List<ImportLogDTO> importPartner(MultipartFile multipartFile, boolean skipHeader) throws Exception {
 
-        File file = Utils.convertMultiPartToFile(multipartFile);
-
         List<ImportLogDTO> importLogs = new ArrayList<ImportLogDTO>();
         List<BusinessPartner> partners = new ArrayList<>();
-
+        File file = null;
         try {
+            file = uploadFileService.saveToUpload(multipartFile);
             String encoding = UniversalDetector.detectCharset(file);
             if(encoding == null) {
                 encoding = "Shift-JIS";
@@ -247,8 +256,13 @@ public class BusinessPartnerService {
                 }
             }
         } catch (Exception e) {
+            logger.error(ExceptionUtils.getStackTrace(e));
+            if(file != null) {
+                file.delete();
+            }
             throw new Exception("取引先のインポートに失敗しました");
-        } finally {
+        }
+        if(file != null) {
             file.delete();
         }
 
@@ -257,12 +271,13 @@ public class BusinessPartnerService {
 
     public List<ImportLogDTO> importPartnerGroup(MultipartFile multipartFile, boolean skipHeader) throws Exception {
 
-        File file = Utils.convertMultiPartToFile(multipartFile);
+        File file = null;
 
         List<ImportLogDTO> importLogs = new ArrayList<ImportLogDTO>();
         List<CSVPartnerGroupDTO> partnerGroups = new ArrayList<>();
 
         try {
+            file = uploadFileService.saveToUpload(multipartFile);
             String encoding = UniversalDetector.detectCharset(file);
             if(encoding == null) {
                 encoding = "Shift-JIS";
@@ -335,8 +350,13 @@ public class BusinessPartnerService {
                 }
             }
         } catch (Exception e) {
+            logger.error(ExceptionUtils.getStackTrace(e));
+            if(file != null) {
+                file.delete();
+            }
             throw new Exception("取引先グループのインポートに失敗しました");
-        } finally {
+        }
+        if(file != null) {
             file.delete();
         }
